use python 3.8 no need for conftest at the moment relax test a bit
